services/agent-monitor/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
watchdog>=4.0.0
websockets>=12.0
aiofiles>=23.2.0
python-frontmatter>=1.1.0
pydantic>=2.0.0
pexpect>=4.9.0
